bug fixed

+++ src/fitcache_client.cpp
@@ -136,7 +136,7 @@ bool fitcache_track_file(const char *path, int flags, int fd)
 			g_mercury_init = true;
 		}
 		// Decide which server should we sent data
-		int host = std::hash<std::string>{}(fd_map[fd]) % g_fitcache_server_count;
+		uint32_t host = std::hash<std::string>{}(fd_map[fd]) % g_fitcache_server_count;
 		L4C_INFO("Remote open - Host %d", host);
 		fitcache_client_comm_gen_open_rpc(host, fd_map[fd], fd);
 
@@ -162,7 +162,7 @@ ssize_t fitcache_remote_read(int fd, void *buf, size_t count)
 	 */
 	ssize_t bytes_read = -1;
 	if (fitcache_file_tracked(fd)){
-		int host = std::hash<std::string>{}(fd_map[fd]) % g_fitcache_server_count;			// The host is the same host when open the file in fd_map[fd]
+		uint32_t host = std::hash<std::string>{}(fd_map[fd]) % g_fitcache_server_count;			// The host is the same host when open the file in fd_map[fd]
 		L4C_INFO("Remote read - Host %d", host);		
 		fitcache_client_comm_gen_read_rpc(host, fd, buf, count, -1);
 		bytes_read = fitcache_read_block_for_file(fd_map[fd]);   		
@@ -187,7 +187,7 @@ ssize_t fitcache_remote_pread(int fd, void *buf, size_t count, off_t offset)
 	 */
 	ssize_t bytes_read = -1;
 	if (fitcache_file_tracked(fd) && fd_redir_map[fd] != 0){
-		int host = std::hash<std::string>{}(fd_map[fd]) % g_fitcache_server_count;	
+		uint32_t host = std::hash<std::string>{}(fd_map[fd]) % g_fitcache_server_count;	
 		L4C_INFO("Remote pread - Host %d", host);		
 		fitcache_client_comm_gen_read_rpc(host, fd, buf, count, offset);
 		bytes_read = fitcache_read_block_for_file(fd_map[fd]);   	
@@ -196,7 +196,7 @@ ssize_t fitcache_remote_pread(int fd, void *buf, size_t count, off_t offset)
 	return bytes_read;
 }
 
-ssize_t fitcache_remote_lseek(int fd, int offset, int whence)
+ssize_t fitcache_remote_lseek(int fd, int64_t offset, int whence)
 {
     FitCache_TIMING("HvacClient_remote_lseek_total");
     
@@ -207,7 +207,7 @@ ssize_t fitcache_remote_lseek(int fd, int offset, int whence)
 	 */
 	ssize_t bytes_read = -1;
 	if (fitcache_file_tracked(fd)){
-		int host = std::hash<std::string>{}(fd_map[fd]) % g_fitcache_server_count;	
+		uint32_t host = std::hash<std::string>{}(fd_map[fd]) % g_fitcache_server_count;	
 		// L4C_INFO("Remote seek - Host %d", host);		
 		fitcache_client_comm_gen_seek_rpc(host, fd, offset, whence);
 		bytes_read = fitcache_read_block_for_file(fd_map[fd]);  // Reuse read_block for seek result		
@@ -219,7 +219,7 @@ ssize_t fitcache_remote_lseek(int fd, int offset, int whence)
 
 void fitcache_remote_close(int fd){
 	if (fitcache_file_tracked(fd)){
-		int host = std::hash<std::string>{}(fd_map[fd]) % g_fitcache_server_count;
+		uint32_t host = std::hash<std::string>{}(fd_map[fd]) % g_fitcache_server_count;
 		fitcache_client_comm_gen_close_rpc(host, fd);             	
 	}
 }
