+++ src/fitcache_internal.h
@@ -141,7 +141,7 @@ extern "C" const char * fitcache_get_path(int fd);
 extern "C" bool  fitcache_remove_fd(int fd);
 extern "C" ssize_t fitcache_remote_read(int fd, void *buf, size_t count);
 extern "C" ssize_t fitcache_remote_pread(int fd, void *buf, size_t count, off_t offset);
-extern "C" ssize_t fitcache_remote_lseek(int fd, int offset, int whence);
+extern "C" ssize_t fitcache_remote_lseek(int fd, int64_t offset, int whence);
 extern "C" void fitcache_remote_close(int fd);
 extern "C" bool fitcache_file_tracked(int fd);
 #endif
@@ -151,7 +151,7 @@ extern const char * fitcache_get_path(int fd);
 extern bool  fitcache_remove_fd(int fd);
 extern ssize_t fitcache_remote_read(int fd, void *buf, size_t count);
 extern ssize_t fitcache_remote_pread(int fd, void *buf, size_t count, off_t offset);
-extern ssize_t fitcache_remote_lseek(int fd, int offset, int whence);
+extern ssize_t fitcache_remote_lseek(int fd, int64_t offset, int whence);
 extern void fitcache_remote_close(int fd);
 extern bool fitcache_file_tracked(int fd);
